labs/02-langchain/demo-11-automating-customer-support-responses/pyproject.toml
[project]
name = "demo-11-automating-customer-support-responses"
version = "0.1.0"
description = "A FastAPI application that uses Google's Gemini AI to generate customer support responses based on customer emails and issue categories."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn==0.40.0",
    "python-dotenv==1.2.1",
    "langchain==1.2.8",
    "langchain-openai==1.1.7",
]
